[REFACTOR] : 로그인 관련 메서드 LogInManager로 옮기고 파이어베이스가 생성하도록 리팩토링.
