+++ Project/Assets/03.Code/Scripts/LogInSystem/ServerSelectPage.cs
@@ -6,32 +6,33 @@ using UnityEngine.UI;
 
 public class ServerSelectPage : Page
 {
-    [SerializeField] private Button _hanyangButton;
-    [SerializeField] private Button _gaeseongButton;
-    [SerializeField] private string lobbySceneName;
+	[SerializeField] private Button _hanyangButton;
+	[SerializeField] private Button _gaeseongButton;
+	[SerializeField] private string lobbySceneName;
 
-    private void OnEnable()
-    {
-        
-        _hanyangButton.onClick.AddListener(() => SetServerName("Hanyang"));
-        _gaeseongButton.onClick.AddListener(() => SetServerName("Gaeseong"));
-    }
+	private void OnEnable()
+	{
 
-    private void OnDisable()
-    {
-        _hanyangButton.onClick.RemoveAllListeners();
-        _gaeseongButton.onClick.RemoveAllListeners();
-    }
+		_hanyangButton.onClick.AddListener(() => SetServerName("Hanyang"));
+		_gaeseongButton.onClick.AddListener(() => SetServerName("Gaeseong"));
+	}
 
-    private async void SetServerName(string name)
-    {
-        if (await FirebaseManager.Instance.SetServerName(name))
-        {
-            SceneManager.LoadScene(lobbySceneName);
-        }
-        else
-        {
+	private void OnDisable()
+	{
+		_hanyangButton.onClick.RemoveAllListeners();
+		_gaeseongButton.onClick.RemoveAllListeners();
+	}
 
-        }
-    }
+	private async void SetServerName(string name)
+	{
+		if (await FirebaseManager.Instance.LogInManager.SetServerName(name))
+		{
+			Destroy(FirebaseManager.Instance.LogInManager.gameObject);
+			SceneManager.LoadScene(lobbySceneName);
+		}
+		else
+		{
+
+		}
+	}
 }
