+++ Project/Assets/03.Code/Scripts/Managers/FirebaseManager.cs
@@ -11,231 +11,57 @@ using Newtonsoft.Json;
 
 public class FirebaseManager : Singleton<FirebaseManager>
 {
-    public FirebaseAuth Auth { get; private set; }
-    public FirebaseDatabase Database { get; private set; }
-    public FirebaseUser User { get; private set; }
+	public FirebaseAuth Auth { get; private set; }
+	public FirebaseDatabase Database { get; private set; }
+	public FirebaseUser User { get; private set; }
 
-    private DatabaseReference _logInUserRef;
+	public LogInManager LogInManager { get; private set; }
+	[SerializeField] private LogInManager _logInManagerPrefab;
 
-    protected override void Awake()
-    {
-        base.Awake();
-        DontDestroyOnLoad(gameObject);
-    }
+	private DatabaseReference _logInUserRef;
 
-    private async void Start()
-    {
-        try
-        {
-            await FirebaseApp.CheckAndFixDependenciesAsync();
+	protected override void Awake()
+	{
+		base.Awake();
+		DontDestroyOnLoad(gameObject);
+	}
 
-            Auth = FirebaseAuth.DefaultInstance;
-            Database = FirebaseDatabase.DefaultInstance;
+	private async void Start()
+	{
+		try
+		{
+			await FirebaseApp.CheckAndFixDependenciesAsync();
 
-            //임시로 로그아웃 했음. 에디터에서 멈추고 해도 그 전 값이 있어서 그럼.
-            //추후에 그냥 구글 연동이나 게스트 로그인은 로그인 화면을 거치지 않고 바로 갈꺼니 지우기만 하면 됌.
-            SignOut();
+			Auth = FirebaseAuth.DefaultInstance;
+			Database = FirebaseDatabase.DefaultInstance;
 
-            Auth.StateChanged += AuthStateChanged;
+			//임시로 로그아웃 했음. 에디터에서 멈추고 해도 그 전 값이 있어서 그럼.
+			//추후에 그냥 구글 연동이나 게스트 로그인은 로그인 화면을 거치지 않고 바로 갈꺼니 지우기만 하면 됌.
+			SignOut();
 
-            print($"현재 사용자: {Auth.CurrentUser?.Email ?? "로그아웃 상태"}");
-        }
-        catch (Exception e)
-        {
-            print($"파이어베이스 초기화 에러 : {e.Message}");
-        }
-    }
+			Auth.StateChanged += AuthStateChanged;
 
-    private void AuthStateChanged(object sender, EventArgs e)
-    {
-        FirebaseAuth senderAuth = sender as FirebaseAuth;
-        if (senderAuth != null)
-            User = senderAuth.CurrentUser; 
-    }
+			LogInManager = Instantiate(_logInManagerPrefab, transform);
 
-    //회원가입
-    public async Task<bool> Create(string email, string password)
-    {
-        try
-        {
-            AuthResult result = await Auth.CreateUserWithEmailAndPasswordAsync(email, password);
+			print($"현재 사용자: {Auth.CurrentUser?.Email ?? "로그아웃 상태"}");
+		}
+		catch (Exception e)
+		{
+			print($"파이어베이스 초기화 에러 : {e.Message}");
+		}
+	}
 
-            LogInUserData userData = new LogInUserData(result.User.UserId);
-            string json = JsonConvert.SerializeObject(userData);
+	private void AuthStateChanged(object sender, EventArgs e)
+	{
+		FirebaseAuth senderAuth = sender as FirebaseAuth;
+		if (senderAuth != null)
+			User = senderAuth.CurrentUser;
+	}
 
-            _logInUserRef = Database.GetReference("loginusers");
-            await _logInUserRef.Child(result.User.UserId).SetRawJsonValueAsync(json);
-            return true;
-        }
-        catch (Exception e)
-        {
-            print($"계정 생성 에러 : {e.Message}");
-
-            return false;
-        }
-    }
-
-    //로그인
-    public async Task<bool> SignIn(string email, string password)
-    {
-        try
-        {
-            AuthResult result = await Auth.SignInWithEmailAndPasswordAsync(email, password);
-
-            string logInTime = DateTime.UtcNow.ToString("o");
-
-            _logInUserRef = Database.GetReference("loginusers");
-
-            try
-            {
-                DataSnapshot snapshot = await _logInUserRef.Child(result.User.UserId).GetValueAsync();
-                LogInUserData userData;
-                if (snapshot.Exists)
-                {
-                    userData = JsonConvert.DeserializeObject<LogInUserData>(snapshot.GetRawJsonValue());
-                    userData.timestamp = logInTime;
-                }
-                else
-                {
-                    userData = new LogInUserData(result.User.UserId, logInTime);
-                }
-
-                string json = JsonConvert.SerializeObject(userData);
-                await _logInUserRef.Child(result.User.UserId).SetRawJsonValueAsync(json);
-
-                Debug.Log($"로그인 성공! 로그인 시간 업데이트: {logInTime}");
-                return true;
-            }
-            catch (Exception e)
-            {
-                print($"데이터를 가져올 수 없음 : {e.Message}");
-
-                //로그인은 성공했는데 데이터를 못가져 왔으니 로그아웃으로 로그인 못하게 해야 함.
-                SignOut();
-                return false;
-            }
-        }
-        catch (Exception e)
-        {
-            print($"로그인 에러 : {e.Message}");
-
-            return false;
-        }
-    }
-
-    //로그아웃
-    public void SignOut()
-    {
-        Auth.SignOut();
-    }
-
-    public async Task<bool> SetNicknameAndRace(string nickname, string race)
-    {
-        try
-        {
-            DataSnapshot snapshot = await _logInUserRef.Child($"{User.UserId}").GetValueAsync();
-            LogInUserData userData;
-
-            if (snapshot.Exists)
-            {
-                userData = JsonConvert.DeserializeObject<LogInUserData>(snapshot.GetRawJsonValue());
-                userData.nickname = nickname;
-                userData.race = race;
-                userData.setNicknameRace = true;
-            }
-            else
-            {
-                userData = new LogInUserData(User.UserId, nickname: nickname, race: race, setNicknameRace: true);
-            }
-
-            string json = JsonConvert.SerializeObject(userData);
-            await _logInUserRef.Child($"{User.UserId}").SetRawJsonValueAsync(json);
-            return true;
-        }
-        catch (Exception e)
-        {
-            print(e.Message);
-            return false;
-        }
-    }
-
-    public async Task<bool> SetServerName(string name)
-    {
-        try
-        {
-            DataSnapshot snapshot = await _logInUserRef.Child($"{User.UserId}").GetValueAsync();
-            LogInUserData userData;
-
-            if (snapshot.Exists)
-            {
-                userData = JsonConvert.DeserializeObject<LogInUserData>(snapshot.GetRawJsonValue());
-                userData.serverName = name;
-            }
-            else
-            {
-                userData = new LogInUserData(User.UserId, serverName: name);
-            }
-
-            string json = JsonConvert.SerializeObject(userData);
-            await _logInUserRef.Child($"{User.UserId}").SetRawJsonValueAsync(json);
-            Debug.Log($"선택한 서버! : {name}");
-            return true;
-        }
-        catch (Exception e)
-        {
-            print($"서버를 선택할 수 없습니다. : {e.Message}");
-            return false;
-        }
-    }
-
-    public async Task<bool> DuplicateNicknameCheck(string nickname)
-    {
-        try
-        {
-            DataSnapshot snapshot = await _logInUserRef.GetValueAsync();
-
-            if (snapshot.Exists)
-            {
-                foreach (DataSnapshot userSnapshot in snapshot.Children)
-                {
-                    if (userSnapshot.HasChild("nickname"))
-                    {
-                        string userNickname = userSnapshot.Child("nickname").Value?.ToString();
-                        if (userNickname.Equals(nickname))
-                        {
-                            return true;
-                        }
-                    }
-                }
-            }
-            return false;
-        }
-        catch (Exception e)
-        {
-            print($"{e.Message}");
-            return true;
-        }
-    }
-
-    public async Task<bool> ExistNicknameAndRace()
-    {
-        try
-        {
-            DataSnapshot snapshot = await _logInUserRef.Child(User.UserId).GetValueAsync();
-            if (snapshot.Exists)
-            {
-                if (snapshot.HasChild("setNicknameRace"))
-                {
-                    bool value = (bool)snapshot.Child("setNicknameRace").Value;
-                    return value;
-                }
-            }
-            return false;
-        }
-        catch (Exception e)
-        {
-            print(e.Message);
-            return false;
-        }
-    }
+	//로그아웃
+	public void SignOut()
+	{
+		//인게임에서 사용할 로그아웃
+		Auth.SignOut();
+	}
 }
