+++ Project/Assets/04.Level/Prefabs/SEO/LogInManager.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &5625287917870097019
+--- !u!1 &247331880682941660
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,40 +8,39 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 7548304237863882960}
-  - component: {fileID: 8388220840350606565}
+  - component: {fileID: 2171300199233058834}
+  - component: {fileID: 181660665155757417}
   m_Layer: 0
-  m_Name: FirebaseManager
+  m_Name: LogInManager
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &7548304237863882960
+--- !u!4 &2171300199233058834
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5625287917870097019}
+  m_GameObject: {fileID: 247331880682941660}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalPosition: {x: 7.721105, y: 8.619133, z: -70.790504}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &8388220840350606565
+--- !u!114 &181660665155757417
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5625287917870097019}
+  m_GameObject: {fileID: 247331880682941660}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 14324499038077341a34bd3a9fa5655e, type: 3}
+  m_Script: {fileID: 11500000, guid: 795d0c6b21a444b4399b074e513ac04d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _logInManagerPrefab: {fileID: 181660665155757417, guid: 49e7920cb0e0bb8428bbed594c7071f1, type: 3}
