+++ Project/Assets/04.Level/Prefabs/SEO/LogInManager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 49e7920cb0e0bb8428bbed594c7071f1
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
